Change mc children's look when delivered
Normally children's appearance is exactly the same as their parents.
This update add a HeroCreatorPatch.cs and hook into the DeliverOffSpring
Method.

Add a postfix to alter
__result.CharacterObject.StaticBodyPropertiesMins.
Set this value with a random seed to alter his appearance.

The parameter name in harmony patch should be exactly the same as the
original target method.